🚀feat(messagesController) add image file serving ⚙️chore(package.json) add hbs as dependency ⚙️chore(server) set view engine and add views directory 🚀feat(server) add route and template for home page
🚀feat: serve an image in messagesController

Added path module and sendFile() method to getMessages function, which now serves an image file from the public directory when a GET request is made.

⚙️chore: add Handlebars as a dependency

This project now has Handlebars as a template engine dependency that can be used for server-side rendering. Handlebars is added to the "dependencies" object in package.json.

⚙️chore: set view engine and add views directory

Server.js was updated with 2 statements that specify the view engine and the directory in which the view files will be stored ready for rendering.

🚀feat: add route and template for home page

A GET request to the root URL now returns a rendered HTML page with a dynamic title and caption and an image sourced from the public directory.